CSV import roundtrip
This change is a temporary fix designed to allow Descriptions to be
exported from an AtoM system and reimported to the SAME AtoM system
when one of the fields 'identifier', 'title' or 'reponame' has been
modified in the CSV.

Normally in this circumstance keymap matching will fail, and secondary
matching on title, identifier and repo name will also fail.

This fix add matching logic from legacyId straight to
information_object.id. This fix is unsafe for any other CSV import
scenario.